% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbsr_pdf.R
\name{bbsr_pdf}
\alias{bbsr_pdf}
\title{Rmarkdown bbsr_pdf template function}
\usage{
bbsr_pdf(...)
}
\arguments{
\item{...}{Parameters to pass to pdf_document2().}
}
\description{
See \code{\link[dsreportr]{ds_pdf}} for details.
}
